# Further improve readAll(string) performance