--- a/runtime/include/qio/qio_error.h
+++ b/runtime/include/qio/qio_error.h
@@ -166,6 +166,7 @@ qioerr qio_mkerror_errno(void);
 #define QIO_ENOMEM (qio_int_to_err(ENOMEM))
 #define QIO_ESHORT (qio_int_to_err(ESHORT))
 #define QIO_EEOF (qio_int_to_err(EEOF))
+#define QIO_EILSEQ (qio_int_to_err(EILSEQ))
 
 
 // This could be done optionally only under Chapel.