Follow-up to PR #24783.

This PR is intended to improve the performance of `readAll(string)` and other functions using `readStringImpl`. It adds `qio_channel_read_chars` which is similar to `qio_channel_read_char`, but can read many codepoints at once, and then uses it in `readStringImpl`.

How does it affect performance? These measurements were collected on my PC with `start_test -performance test/studies/shootout/submitted/regexredux2.chpl --numtrials 4` (and note that `main` here includes #24783). Please note that the performance gain for `readAll` itself will be a larger percentage (since it is not most of the benchmark time). The speed gain for `readAll(string)` specifically within the benchmark is about 2x.


| Benchmark    | main   | this PR | speedup    |
| ------------ | ------ | ------- | ---------- |
| regexredux2  | 1.42 s | 1.37 s  |  5% faster |


Reviewed by @jeremiah-corrado - thanks!

- [x] full comm=none testing