@mppf 2024-04-12
I've got these failures to look at

```
[Error matching compiler output for chpldoc/compflags/combinations/version.doc]
[Error matching compiler output for chpldoc/compflags/combinations/z_copy_version.doc]
[Error matching compiler output for chpldoc/compflags/combinations/z_copyright_help_version.doc]
[Error matching compiler output for chpldoc/compflags/combinations/z_help_version.doc]
[Error matching program output for library/standard/IO/advanceThrough/multiByteNotPresent]
[Error matching program output for library/standard/IO/advanceTo/multiByteNotPresent]
[Error matching program output for regex/ferguson/rechan]
[Error matching program output for regex/empty-channel-matches]
[Error matching program output for release/examples/benchmarks/shootout/revcomp-fast]
[Error matching program output for studies/shootout/reverse-complement/bharshbarg/revcomp-mark-skip-read-begin]
```